refactor: separate timeline for utils
